requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-leaflet==0.31.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
